patch: Only clean up the Renderer object if it is a CanvasRenderer, as WebGL Renderer does not need to be deleted
Original issue: https://github.com/rive-app/rive-wasm/issues/333

Only the canvas renderer can have its emscripten object deleted on cleanup, but not the WebGL one. So just adding a quick check to only delete if it detects the runtime has a canvas renderer.

Diffs=
c579540d1 patch: Only clean up the Renderer object if it is a CanvasRenderer, as WebGL Renderer does not need to be deleted (#6037)

+++ js/src/rive.ts
@@ -1678,8 +1678,11 @@ export class Rive {
     this.stopRendering();
     // Clean up any artboard, animation or state machine instances.
     this.cleanupInstances();
-    // Delete the renderer
-    this.renderer?.delete();
+    // Only CanvasRenderer needs a delete for now, not WebGL.. might be a better
+    // way to detect this in the future 
+    if (this.runtime?.CanvasRenderer) {
+      this.renderer?.delete();
+    }
     this.renderer = null;
     // Delete the rive file
     this.file?.delete();
